rust7::client

Struct S7Client

pub struct S7Client {
    pub pdu_length: u16,
    pub connected: bool,
    pub last_time: f64,
    pub chunks: usize,
    /* private fields */
}

Fields
pdu_length: u16

PDU length negotiated by the CPU

connected: bool

Client connected

last_time: f64

Last Job time (ms).

If an error occurred the value will be 0

chunks: usize

Indicates how many pieces the data to be read or written in the last operation was divided into

Maybe you need to know it only for extreme tuning

Implementations

impl S7Client

pub fn new() -> Self

Creates a new S7Client instance with default settings.

The client starts disconnected, use one of connect_XXX methods to open a connection to a PLC.

Returns

A new S7Client ready to connect.

pub fn set_connection_type(&mut self, connection_type: u16)

Changes the S7 connection type to the PLC

The three possible connection types are:

CT_PG: (as a programming device)

CT_OP: (as an HMI)

CT_S7: (as a generic device)

In practice, there aren’t many differences; the S7_PG connection should ensure better system responsiveness, but in reality, I’ve never noticed any noticeable differences.

CT_PG is used by default.

With very old PLCs (early S7300 series) that have limited communication resources, the connection may be rejected if we have S7Manager with many online windows open at the same time. In this case, use S7_OP or S7_BASIC.

Parameters

connection_type: Connection type.

Notes

The client must not be connected (that is, call this method before connecting).

This method is not useful if you use connect_tsap() because the connection_type is already contained in the REMOTE_TSAP record.

pub fn set_timeout( &mut self, co_timeout_ms: u64, rd_timeout_ms: u64, wr_timeout_ms: u64, )

Sets operations timeout

Parameters

co_timeout_ms : TCP Connection timeout (ms) (Default = 3000 ms)

rd_timeout_ms : Read Connection timeout (ms) (Default = 1000 ms)

wr_timeout_ms : Write Connection timeout (ms) (Default = 500 ms)

Notes

Values must be > 0, otherwise they are ignored

The client must not be connected (that is, call this method before connecting).

pub fn set_connection_port(&mut self, port: u16)

Sets the TCP Connection Port

The default S7 Port is 102, but if you need NAT the addresses you can use this method to change the default value.

Parameters

port: TCP Connection port (1..65535)

Notes

Value must be > 0, otherwise it is ignored

The client must not be connected (that is, call this method before connecting).

pub fn connect_s71200_1500(&mut self, ip: &str) -> Result<(), S7Error>

Connects to the S71200 or S71500 families

This helper method is same as connect_rack_slot() with rack=0 and slot=0

Parameters

ip : PLC IPV4 address.

For Notes, Return and Errors look at connect_tsap()

pub fn connect_s7300(&mut self, ip: &str) -> Result<(), S7Error>

Connects to the S7300 family

This helper method is same as connect_rack_slot() with rack=0 and slot=2

Parameters

ip : PLC IPV4 address.

For Notes, Return and Errors look at connect_tsap()

pub fn connect_rack_slot( &mut self, ip: &str, rack: u16, slot: u16, ) -> Result<(), S7Error>

Connects to a Siemens PLC/Drive using Rack and Slot

Rack and Slot are Hardware configuration parameters.

For S7300 and S71200/1500 they are fixed, (see connect_s7300() and connect_s71200_1500() ).

Ultimately, you will need of this method only to connect to S7400, WinAC or other Siemens hardware, like Drives, which Rack and Slot can vary.

Parameters

ip : PLC IPV4 address.

rack : CPU/CU Rack.

slot : CPU/CU Slot.

For Notes, Return and Errors look at connect_tsap()

pub fn connect_tsap( &mut self, ip: &str, local_tsap: u16, remote_tsap: u16, ) -> Result<(), S7Error>

Connects to a Siemens ISO-Hardware using TSAP records

This is the deepest connection method, you will need it only to connect to LOGO! or S7200. It’s internally called by all other connection methods.

Parameters

ip : PLC IPV4 address.

local_tsap : Client TSAP.

remote_tsap : Server TSAP (PLC).

Notes

The connection port used is 102 (S7Protocol Port) unless you
changed it via set_connection_port()

Returns

Ok(()) on success, or an S7Error on failure.

Errors

S7Error::TcpConnectionFailed: TCP connection could not be established.

S7Error::IsoConnectionFailed: ISO connection failed

S7Error::PduNegotiationFailed: PDU negotiation failed.

S7Error::Io: network I/O error.

pub fn disconnect(&mut self)

Closes the connection.

Safe to call even if the client is not currently connected. After disconnection, calls to read/write will return S7Error::NotConnected.

Notes

A Client should be disconnected on low-level error (see read_area() and write_area() suggestion)

pub fn read_area( &mut self, area: u8, db_number: u16, start: u16, wordlen: u8, buffer: &mut [u8], ) -> Result<(), S7Error>

Reads a block of data from a specific S7 memory area.

Parameters

area: S7 memory area constant (e.g., S7_AREA_PE, S7_AREA_PA, S7_AREA_DB, S7_AREA_MK).

db_number: DB number (ignored for non-DB areas).

start: Starting element index (byte index for bytes, bit index for bits).

wordlen: Word length constant (e.g., S7_WL_BYTE, S7_WL_BIT).

buffer: Destination buffer to store the read data.

Values

area

S7_AREA_PE (0x81): Process Inputs

S7_AREA_PA (0x82): Process Outputs

S7_AREA_MK (0x83): Merkers

S7_AREA_DB (0x84): Data Block

wordlen

S7_WL_BIT (0x01) : Bit access

S7_WL_BYTE (0x02): Byte access

Bit access notes

The start must be expressed in bits. For example, if you want to access bit DBX 45.3, the start value would be 45 * 8 + 3 = 363.

Whatever buffer is passed, only the first byte will be used, which is considered true if !=0 or false if ==0

Returns

Ok(()) Operation succeeded.

Errors

Low level

S7Error::NotConnected: An attempt was made to read while the client was not connected.

S7Error::IsoInvalidHeader: Invalid ISO Header

S7Error::IsoInvalidTelegram: Inconsistent expected telegram length.

S7Error::IsoFragmentedPacket: ISO Packet fragmented.

S7Error::S7Unspecified: Unknown S7 Error.

S7Error::Io: network I/O error.

Suggestion

In case of a low-level error, it is highly recommended to disconnect and reconnect the Client (as WinCC or other SCADA do)

High level

S7Error::NotFound: The resource was not found (e.g. Inexistent DB).

S7Error::S7InvalidAddress:

Attempt to read beyond the limits.

The DB is optimized.

Notes

The number of bytes to read will be equal to the size of the buffer passed.

Large blocks are automatically split into chunks based on the negotiated PDU size.

In case of error the buffer contents will be inconsistent and should not be considered.

pub fn write_area( &mut self, area: u8, db_number: u16, start: u16, wordlen: u8, buffer: &[u8], ) -> Result<(), S7Error>

Writes a block of data to a specific S7 memory area.

Parameters

area: S7 memory area constant (e.g., S7_AREA_PE, S7_AREA_PA, S7_AREA_DB, S7_AREA_MK).

db_number: DB number (ignored for non-DB areas).

start: Starting element index (byte index for bytes, bit index for bits).

wordlen: Word length constant (e.g., S7_WL_BYTE, S7_WL_BIT).

buffer: Source buffer to write.

Values

area

S7_AREA_PE (0x81): Process Inputs

S7_AREA_PA (0x82): Process Outputs

S7_AREA_MK (0x83): Merkers

S7_AREA_DB (0x84): Data Block

wordlen

S7_WL_BIT (0x01) : Bit access

S7_WL_BYTE (0x02): Byte access

Bit access notes

The start must be expressed in bits. For example, if you want to access bit DBX 45.3, the start value would be 45 * 8 + 3 = 363.

Whatever buffer is passed, only the first byte will be used, which is considered true if !=0 or false if ==0

Writing a bit affects only that bit, leaving adjacent bits in the byte unchanged.

Returns

Ok(()) Operation succeeded.

Errors

Low level

S7Error::NotConnected: An attempt was made to write while the client was not connected.

S7Error::IsoInvalidHeader: Invalid ISO Header

S7Error::IsoInvalidTelegram: Inconsistent expected telegram length.

S7Error::IsoFragmentedPacket: ISO Packet fragmented.

S7Error::S7Unspecified: Unknown S7 Error.

S7Error::Io: network I/O error.

Suggestion

In case of a low-level error, it is highly recommended to disconnect and reconnect the Client (as WinCC or other SCADA do)

High level

S7Error::NotFound: The resource was not found (e.g. Inexistent DB).

S7Error::S7InvalidAddress:

Attempt to write beyond the limits.

The DB is optimized.

Notes

The number of bytes to write will be equal to the size of the buffer passed.

Large blocks are automatically split into chunks based on the negotiated PDU size.

Writing the output buffer (S7_AREA_PA) usually does not produce useful results, in fact the output process image will be rewritten by OB1 in the next round

pub fn read_db( &mut self, db_number: u16, start: u16, buffer: &mut [u8], ) -> Result<(), S7Error>

Reads a block of byte from a specific Data Block (DB)

This helper method is same as read_area() with:

area = S7_AREA_DB

wordlen = S7_WL_BYTE

Parameters

db_number: DB number

start: Starting byte index

buffer: Destination buffer to store the read data.

Notes

The number of bytes to read will be equal to the size of the buffer passed.

For further info, please refer to read_area()

pub fn read_bit( &mut self, area: u8, db_number: u16, byte_num: u16, bit_idx: u8, ) -> Result<bool, S7Error>

Reads a bit from a specific S7 memory area

This helper method is same as read_area() with:

wordlen = S7_WL_BIT

start = byte_num * 8 + bit_idx

Parameters

area: S7 memory area constant (e.g., S7_AREA_PE, S7_AREA_PA, S7_AREA_DB, S7_AREA_MK).

db_number: DB number (ignored for non-DB areas).

byte_num: Byte Number.

bit_idx: Bit index inside the byte (0..7).

Example

To read DB10.DBX71.4 use:

my_bit = read_bit(S7_AREA_DB, 10, 71, 4);

Returns

Ok(<bool>) or Err(<S7Error>)

Suggestion

Even reading a single bit requires an entire telegram.
Since reading is non-invasive, if you need to read multiple bits 
(more or less adjacent in the same area), I recommend reading blocks 
of bytes and then unpacking them.

For further info, please refer to read_area()

pub fn write_db( &mut self, db_number: u16, start: u16, buffer: &[u8], ) -> Result<(), S7Error>

Writes a block of byte to a specific Data Block (DB)

This helper method is same as write_area() with:

area = S7_AREA_DB

wordlen = S7_WL_BYTE

Parameters

db_number: DB number

start: Starting byte index

buffer: Source buffer to write.

Notes

The number of bytes to write will be equal to the size of the buffer passed.

For further info, please refer to write_area()

pub fn write_bit( &mut self, area: u8, db_number: u16, byte_num: u16, bit_idx: u8, value: bool, ) -> Result<(), S7Error>

Writes a bit to a specific S7 memory area

This helper method is same as write_area() with:

wordlen = S7_WL_BIT

start = byte_num * 8 + bit_idx

Parameters

area: S7 memory area constant (e.g., S7_AREA_PE, S7_AREA_PA, S7_AREA_DB, S7_AREA_MK).

db_number: DB number (ignored for non-DB areas).

byte_num: Byte Number.

bit_idx: Bit index inside the byte (0..7).

value: Value to write (true | false).

Example

To write 1 into DB10.DBX71.4 use:

write_bit(S7_AREA_DB, 10, 71, 4, true);

Returns

Ok(()) Operation succeeded.

Notes

Writing a bit affects only that bit, leaving adjacent bits in the byte unchanged. 

For further info, please refer to write_area()